+++ src/global/variables/RandomClause.java
@@ -1,4 +1,4 @@
-package global.variables.clauses;
+package global.variables;
 
 public interface RandomClause<T> {
     void trackValue(int groupNum, String newListValue);
