+++ src/global/variables/clauses/RandomChar.java
@@ -1,6 +1,7 @@
 package global.variables.clauses;
 
 import global.variables.Clause;
+import global.variables.RandomClause;
 
 import java.util.ArrayList;
 import java.util.HashMap;
