+++ src/global/variables/clauses/RandomDouble.java
@@ -3,6 +3,7 @@ package global.variables.clauses;
 import global.utils.RandomUtil;
 import global.utils.RegexUtil;
 import global.variables.Clause;
+import global.variables.RandomClause;
 
 import java.util.ArrayList;
 import java.util.HashMap;
